Add advanced search function
Backend advanced search is available on the api using DCI query language.
Add this feature on the UI with a shortcut for now.
ctrl + q iq the shortcut to display or hide the search entry using the query language on the ui.

search is only available on the jobs page.

Change-Id: I154dc829d6d438eeca2601a200fd779c6391c717