Introduce the new team creation wizard
The team creation is always attached to the creation of users and permissions.
We could improve the user experience by setting everything in a form wizard.

We can create the team, users, attach the users to the team and set the permissions to the team in one task.

The user creation should get or create the user if it exists, and attach it to the team. It should display a good error message if an error occurs.

A user created only with the Red Hat login, should have an empty password. It will be using his SSO account.

Depends-On: Ibc7360594e540063b83ac5dfd04f4e269355d00d
Change-Id: I0d2c5a416e182afc7a433ecbad55fd57d3a8ad21