Fix groupJobByPipeline with jobs keep job order
Job are grouped together in reverse order because children jobs are coming after the parent job.
But the order of the first level is lost because of the hashmap used to group the jobs together.
This patch fixes the issue by keeping the order of the first level of the job array.

Change-Id: Ib320fb87e55b5dfb438453d073c35192005bd3a5