Introduce pagination on DCI Analytics
pagination is needed in some analytics pages like Test Analysis because we can't grab all jobs.
This patch introduce pagination.

We also change the way we save the queries. Now we are saving the complete search url and pathname.
We migrate the old query with the new one.
This patch prepare the queries saved on the control server.

Change-Id: I1deac31a699710224e585e33f522760ac68beab1