Clarify Team Roles and Improve UI Navigation for Administrative Operations (#9)
Implementation of ADR 001

Currently, users can belong to multiple teams. For example the admin team, the EPMs team, the Red Hat team and a regular team. The distinction between a regular team and an administrative team (EPM or Admin) is unclear in the UI. Users report confusion, for example, when they cannot see components in certain topics or encounter 401 errors after switching teams. Additionally, Remoteci creation does not clearly associate with the correct team, and the “EPMs” team name does not accurately reflect its moderation role.